+++ Cargo.toml
@@ -5,8 +5,9 @@ members = [
     "pijama_ast",
     "pijama_parser",
     "pijama_ty",
-    "pijama_mir",
+    "pijama_hir",
     "pijama_tycheck",
+    "pijama_mir",
     "pijama_lir",
     "pijama_machine",
     "pijama_driver",
