+++ pijama_hir/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "pijama_tycheck"
+name = "pijama_hir"
 version = "0.1.0"
 authors = ["The Pijama Project Developers"]
 edition = "2018"
 
 [dependencies]
 thiserror = "1.0"
-pijama_hir = {path = "../pijama_hir"}
+pijama_ast = {path = "../pijama_ast"}
 pijama_ty = {path = "../pijama_ty"}
-pijama_common = {path = "../pijama_common"}
+pijama_common = { path = "../pijama_common" }
